port2: Fix _Bool truncation bugs for variables storing non-boolean values
C99 _Bool truncates all non-zero values to 1, breaking variables that
the original K&R code (typedef short bool) used to store wider values:
- escape/aescaped: stored '\n' (10), making Enter terminate insert mode
- inopen: tri-state (-1, 0, 1) for undo logic
- inglobal: tri-state (0, 1, 2) for global-within-global detection
- gobblebl: countdown (3→0) for wrapmargin space gobbling
- pfast: 2-bit bitfield (0–3) for terminal output mode

Changed all five from bool to int, updated prototypes, and removed
now-unnecessary (int) casts.

+++ port2/ex_cmds.c
@@ -797,7 +797,7 @@ numberit:
 		case '=':
 			donewline();
 			setall();
-			if ((int)inglobal == 2)
+			if (inglobal == 2)
 				pofix();
 			printf("%d", lineno(addr2));
 			noonl();
